Rename type check folder to semantics
